# src/main/java/com/itmo/blse/a12n/dto/UserRegistrationRequest.java
package com.itmo.blse.a12n.dto;

import com.itmo.blse.tournaments.model.Roles;
import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

import java.util.List;

@Getter
@Setter
@Builder
public class UserRegistrationRequest {

    String username;
    String password;
    List<Roles> roles;

}